renamed ClientProxy.ts to clientProxy.ts for #85
